% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.firehose_package.R
\docType{package}
\name{paws.firehose-package}
\alias{paws.firehose}
\alias{paws.firehose-package}
\title{paws.firehose: Amazon Kinesis Firehose}
\description{
Amazon Kinesis Data Firehose is a fully managed
service that delivers real-time streaming data to destinations such as
Amazon Simple Storage Service (Amazon S3), Amazon Elasticsearch
Service (Amazon ES), Amazon Redshift, and Splunk.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker
}

}
\keyword{internal}
