% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.firehose_operations.R
\name{untag_delivery_stream}
\alias{untag_delivery_stream}
\title{Removes tags from the specified delivery stream}
\usage{
untag_delivery_stream(
  DeliveryStreamName = "string",
  TagKeys = list(
    "string"
  )
)
}
\arguments{
\item{DeliveryStreamName}{[required] The name of the delivery stream.}

\item{TagKeys}{[required] A list of tag keys. Each corresponding tag is removed from the delivery stream.}
}
\description{
Removes tags from the specified delivery stream. Removed tags are deleted, and you can't recover them after this operation successfully completes.
}
\details{
If you specify a tag that doesn't exist, the operation ignores it.

This operation has a limit of five transactions per second per account.
}
\examples{

}
