% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.firehose_operations.R
\name{update_destination}
\alias{update_destination}
\title{Updates the specified destination of the specified delivery stream}
\usage{
update_destination(DeliveryStreamName, CurrentDeliveryStreamVersionId,
  DestinationId, S3DestinationUpdate = NULL,
  ExtendedS3DestinationUpdate = NULL, RedshiftDestinationUpdate = NULL,
  ElasticsearchDestinationUpdate = NULL, SplunkDestinationUpdate = NULL)
}
\arguments{
\item{DeliveryStreamName}{The name of the delivery stream.}

\item{CurrentDeliveryStreamVersionId}{Obtain this value from the \strong{VersionId} result of DeliveryStreamDescription. This value is required, and helps the service perform conditional operations. For example, if there is an interleaving update and this value is null, then the update destination fails. After the update is successful, the \code{VersionId} value is updated. The service then performs a merge of the old configuration with the new configuration.}

\item{DestinationId}{The ID of the destination.}

\item{S3DestinationUpdate}{[Deprecated] Describes an update for a destination in Amazon S3.}

\item{ExtendedS3DestinationUpdate}{Describes an update for a destination in Amazon S3.}

\item{RedshiftDestinationUpdate}{Describes an update for a destination in Amazon Redshift.}

\item{ElasticsearchDestinationUpdate}{Describes an update for a destination in Amazon ES.}

\item{SplunkDestinationUpdate}{Describes an update for a destination in Splunk.}
}
\description{
Updates the specified destination of the specified delivery stream.
}
\details{
Use this operation to change the destination type (for example, to replace the Amazon S3 destination with Amazon Redshift) or change the parameters associated with a destination (for example, to change the bucket name of the Amazon S3 destination). The update might not occur immediately. The target delivery stream remains active while the configurations are updated, so data writes to the delivery stream can continue during this process. The updated configurations are usually effective within a few minutes.

Switching between Amazon ES and other services is not supported. For an Amazon ES destination, you can only update to another Amazon ES destination.

If the destination type is the same, Kinesis Data Firehose merges the configuration parameters specified with the destination configuration that already exists on the delivery stream. If any of the parameters are not specified in the call, the existing values are retained. For example, in the Amazon S3 destination, if EncryptionConfiguration is not specified, then the existing \code{EncryptionConfiguration} is maintained on the destination.

If the destination type is not the same, for example, changing the destination from Amazon S3 to Amazon Redshift, Kinesis Data Firehose does not merge any parameters. In this case, all parameters must be specified.

Kinesis Data Firehose uses \strong{CurrentDeliveryStreamVersionId} to avoid race conditions and conflicting merges. This is a required field, and the service updates the configuration only if the existing configuration has a version ID that matches. After the update is applied successfully, the version ID is updated, and can be retrieved using DescribeDeliveryStream. Use the new version ID to set \strong{CurrentDeliveryStreamVersionId} in the next call.
}
\examples{

}
