% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.firehose_operations.R
\name{list_delivery_streams}
\alias{list_delivery_streams}
\title{Lists your delivery streams}
\usage{
list_delivery_streams(Limit = NULL, DeliveryStreamType = NULL,
  ExclusiveStartDeliveryStreamName = NULL)
}
\arguments{
\item{Limit}{The maximum number of delivery streams to list. The default value is 10.}

\item{DeliveryStreamType}{The delivery stream type. This can be one of the following values:
\itemize{
\item \code{DirectPut}: Provider applications access the delivery stream directly.
\item \code{KinesisStreamAsSource}: The delivery stream uses a Kinesis data stream as a source.
}

This parameter is optional. If this parameter is omitted, delivery streams of all types are returned.}

\item{ExclusiveStartDeliveryStreamName}{The name of the delivery stream to start the list with.}
}
\description{
Lists your delivery streams.
}
\details{
The number of delivery streams might be too large to return using a single call to \code{ListDeliveryStreams}. You can limit the number of delivery streams returned, using the \strong{Limit} parameter. To determine whether there are more delivery streams to list, check the value of \code{HasMoreDeliveryStreams} in the output. If there are more delivery streams to list, you can request them by specifying the name of the last delivery stream returned in the call in the \code{ExclusiveStartDeliveryStreamName} parameter of a subsequent call.
}
\section{Accepted Parameters}{
\preformatted{list_delivery_streams(
  Limit = 123,
  DeliveryStreamType = "DirectPut"|"KinesisStreamAsSource",
  ExclusiveStartDeliveryStreamName = "string"
)
}
}

